refac: consolidate Help functionality into TopRightControls and remove Help component (#100)
* refactor: consolidate Help functionality into TopRightControls and remove Help component

* refactor: improve formatting and readability of Help Button component

* fix: update Help Icon size to match other icons in TopRightControls

* feat: replace Help Icon SVG with QuestionMarkCircle component in TopRightControls

---------
